feat: spmc/mpsc ring buffer inplementations

+++ connector/ring_buffer.go
@@ -7,11 +7,11 @@ import (
 // ErrClosed is returned when the ring buffer is closed.
 var ErrClosed = rb.ErrClosed
 
-func newRingBuffer[T any](capacity uint32) *RingBuffer[T] {
+func newRingBuffer[T any](capacity uint64) *RingBuffer[T] {
 	return rb.NewRingBuffer[T](capacity, rb.BufferKindSPSC)
 }
 
 // NewRingBuffer returns a new lock-free spsc generic ring buffer.
-func NewRingBuffer[T msgBody](capacity uint32) *RingBuffer[*msgWrap[T]] {
+func NewRingBuffer[T msgBody](capacity uint64) *RingBuffer[*msgWrap[T]] {
 	return newRingBuffer[*msgWrap[T]](capacity)
 }
